Create ABC.java
Context behind the code:

You will be given three integers A, B and C. The numbers will not be given in that exact order, but we do know that A is less than B and B less than C. In order to make for a more pleasant viewing, we want to rearrange them in a given order.

The first line of input contains the three positive integers A, B and C, not necessarily in that order. The three numbers will be less than or equal to 100. The second line of input contains three uppercase letters ’A’, ’B’ and ’C’ (with no spaces between them) representing the desired order.

The code outputs A, B, and C in the desired order on a single line, separated by single spaces.